+++ cell/test/u_cell_info_test.c
@@ -102,7 +102,7 @@
 
 /** Used for keepGoingCallback() timeout.
  */
-static int64_t gStopTimeMs;
+static int32_t gStopTimeMs;
 
 /** Handles.
  */
